+++ lib/src/service/tweets/tweet_data.g.dart
@@ -227,6 +227,12 @@ const _$LanguageEnumMap = {
   Language.georgian: 'ka',
   Language.lithuanian: 'lt',
   Language.undefined: 'und',
+  Language.mentionsOnly: 'qam',
+  Language.cashtagsOnly: 'qct',
+  Language.hashtagsOnly: 'qht',
+  Language.mediaLinks: 'qme',
+  Language.shortText: 'qst',
+  Language.mediaOrCardOnly: 'zxx',
 };
 
 const _$ReplySettingEnumMap = {